refactor: extract search into a standalone file
